+++ __main__.py
@@ -1,4 +1,4 @@
-from game.director import Director
+from modules.hilo_director import Director
 
 director = Director()
 director.start_game()
