--- hilo-incomplete/hilo/__main__.py
@@ -1,4 +1,0 @@
-from game.hilo_director import Director
-
-director = Director()
-director.start_game()
